solving some warnings

+++ road_net.cc
@@ -1231,7 +1231,7 @@ pthread_param_short_path* new_pthread_param_short_path
 		unsigned int* pointer,
 		pthread_mutex_t* mutex_pool,
 		RoadNet* net,
-		const unsigned int max_length,
+		const double max_length,
 		 std::vector < std::map < unsigned int , unsigned int > * >* short_paths
 	)
 {
